# Swagger/E2Manager_API.yaml
openapi: 3.0.0
info:
  title: E2 Manager Service
  description: E2 Manager Service APIs
  version: 5.2.21
servers:
  - url: 'http://{apiRoot}/v1'
    variables:
      apiRoot:
        default: 'localhost:3800'
paths:
...
  '/nodeb/enb/{ranName}':
    delete:
      tags:
        - nodeb
      summary: Delete eNB
      operationId: DeleteEnb
      parameters:
        - name: ranName
          in: path
          required: true
          description: Name of RAN to delete
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodebResponse'
        '400':
          description: The specified RAN is not ENB
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: A RAN with the specified name was not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - nodeb
      summary: Update eNB
      operationId: UpdateEnb
      parameters:
        - name: ranName
          in: path
          required: true
          description: Name of ENB ran to update
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEnbRequest'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodebResponse'
        '400':
          description: Invalid input
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  '/nodeb/gnb/{ranName}':
    put:
      summary: Update gNB Cells
      tags:
        - nodeb
      operationId: UpdateGnb
      parameters:
        - name: ranName
          in: path
          required: true
          description: Name of GNB RAN to update
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGnbRequest'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodebResponse'
        '400':
          description: Invalid input
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /nodeb/states:
    get:
      tags:
        - nodeb
      summary: Get RANs identities list
      operationId: getNodebIdList
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NodebIdentity'
        '500':
          description: Internal Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    UpdateGnbRequest:
      type: object
      required:
        - servedNrCells
      properties:
        servedNrCells:
          $ref: '#/components/schemas/ServedNrCells'
      additionalProperties: false
    UpdateEnbRequest:
      type: object
      required:
        - enb
      properties:
        enb:
          $ref: '#/components/schemas/Enb'
      additionalProperties: false
...
    NodebIdentity:
      properties:
        globalNbId:
          properties:
            nbId:
              type: string
            plmnId:
              type: string
          type: object
        inventoryName:
          type: string
      type: object
    ErrorResponse:
      type: object
      required:
        - errorCode
        - errorMessage
      properties:
        errorCode:
          type: string
          description: >-
            401 - corrupted json, 402 - validation error, 403 - RAN in wrong
            state, 404 - resource not found, 500 - RNIB error, 501 - internal
            problem, 502 - RMR error, 503 - Routing Manager Unavailable
        errorMessage:
          type: string
          description: Human readable text
...
    RanFunction:
      properties:
        ranFunctionId:
          type: integer
        ranFunctionDefinition:
          type: string
        ranFunctionRevision:
          type: integer
      type: object
    Enb:
      properties:
        enbType:
          oneOf:
            - type: string
            - type: integer
        guGroupIds:
          items:
            type: string
          type: array
        servedCells:
          items:
            properties:
              bandwidthReducedSi:
                oneOf:
                  - type: string
                  - type: integer
              broadcastPlmns:
                items:
                  type: string
                type: array
              cellId:
                type: string
              choiceEutraMode:
                properties:
                  fdd:
                    properties:
                      dlTransmissionBandwidth:
                        oneOf:
                          - type: string
                          - type: integer
                      dlearFcn:
                        type: integer
                      ulTransmissionBandwidth:
                        oneOf:
                          - type: string
                          - type: integer
                      ulearFcn:
                        type: integer
                    additionalProperties: false
                    type: object
                  tdd:
                    properties:
                      additionalSpecialSubframeExtensionInfo:
                        properties:
                          additionalSpecialSubframePatternsExtension:
                            oneOf:
                              - type: string
                              - type: integer
                          cyclicPrefixDl:
                            oneOf:
                              - type: string
                              - type: integer
                          cyclicPrefixUl:
                            oneOf:
                              - type: string
                              - type: integer
                        additionalProperties: false
                        type: object
                      additionalSpecialSubframeInfo:
                        properties:
                          additionalSpecialSubframePatterns:
                            oneOf:
                              - type: string
                              - type: integer
                          cyclicPrefixDl:
                            oneOf:
                              - type: string
                              - type: integer
                          cyclicPrefixUl:
                            oneOf:
                              - type: string
                              - type: integer
                        additionalProperties: false
                        type: object
                      earFcn:
                        type: integer
                      specialSubframeInfo:
                        properties:
                          cyclicPrefixDl:
                            oneOf:
                              - type: string
                              - type: integer
                          cyclicPrefixUl:
                            oneOf:
                              - type: string
                              - type: integer
                          specialSubframePatterns:
                            oneOf:
                              - type: string
                              - type: integer
                        additionalProperties: false
                        type: object
                      subframeAssignment:
                        oneOf:
                          - type: string
                          - type: integer
                      transmissionBandwidth:
                        oneOf:
                          - type: string
                          - type: integer
                    additionalProperties: false
                    type: object
                additionalProperties: false
                type: object
              csgId:
                type: string
              eutraMode:
                oneOf:
                  - type: string
                  - type: integer
              freqBandIndicatorPriority:
                oneOf:
                  - type: string
                  - type: integer
              mbmsServiceAreaIdentities:
                items:
                  type: string
                type: array
              mbsfnSubframeInfos:
                items:
                  properties:
                    radioframeAllocationOffset:
                      type: integer
                    radioframeAllocationPeriod:
                      oneOf:
                        - type: string
                        - type: integer
                    subframeAllocation:
                      type: string
                    subframeAllocationType:
                      oneOf:
                        - type: string
                        - type: integer
                  additionalProperties: false
                  type: object
                type: array
              multibandInfos:
                items:
                  type: integer
                type: array
              neighbourInfos:
                items:
                  properties:
                    earFcn:
                      type: integer
                    ecgi:
                      type: string
                    pci:
                      type: integer
                    tac:
                      type: string
                  additionalProperties: false
                  type: object
                type: array
              numberOfAntennaPorts:
                oneOf:
                  - type: string
                  - type: integer
              pci:
                type: integer
              prachConfiguration:
                properties:
                  highSpeedFlag:
                    type: boolean
                  prachConfigurationIndex:
                    type: integer
                  prachFrequencyOffset:
                    type: integer
                  rootSequenceIndex:
                    type: integer
                  zeroCorrelationZoneConfiguration:
                    type: integer
                type: object
              tac:
                type: string
              additionalCellInformation:
                $ref: '#/components/schemas/AdditionalCellInformation'
            type: object
          type: array
      type: object
    Gnb:
      properties:
        servedNrCells:
          $ref: '#/components/schemas/ServedNrCells'
        ranFunctions:
          items:
            $ref: '#/components/schemas/RanFunction'
          type: array
      type: object
...
    SetupFailure:
      properties:
        criticalityDiagnostics:
          properties:
            informationElementCriticalityDiagnostics:
              items:
                properties:
                  ieCriticality:
                    oneOf:
                      - type: string
                      - type: integer
                  ieId:
                    type: integer
                  typeOfError:
                    oneOf:
                      - type: string
                      - type: integer
                additionalProperties: false
                type: object
              type: array
            procedureCode:
              type: integer
            procedureCriticality:
              oneOf:
                - type: string
                - type: integer
            triggeringMessage:
              oneOf:
                - type: string
                - type: integer
          additionalProperties: false
          type: object
        miscellaneousCause:
          oneOf:
            - type: string
            - type: integer
        networkLayerCause:
          oneOf:
            - type: string
            - type: integer
        protocolCause:
          oneOf:
            - type: string
            - type: integer
        timeToWait:
          oneOf:
            - type: string
            - type: integer
        transportLayerCause:
          oneOf:
            - type: string
            - type: integer
      additionalProperties: false
      type: object
...
    AdditionalCellInformation:
      properties:
        cellLatitude:
          type: number
        cellLongitude:
          type: number
        antennaHeight:
          type: number
        antennaAzimuthDirection:
          type: number
        antennaTiltAngle:
          type: number
        antennaMaxTransmit:
          type: number
        antennaMaxGain:
          type: number
        sectorId:
          type: integer
      type: object
    NodebResponse:
      properties:
        connectionStatus:
          oneOf:
            - type: string
            - type: integer
        enb:
          $ref: '#/components/schemas/Enb'
        failureType:
          oneOf:
            - type: string
            - type: integer
        globalNbId:
          properties:
            nbId:
              type: string
            plmnId:
              type: string
          type: object
        gnb:
          $ref: '#/components/schemas/Gnb'
        ip:
          type: string
        nodeType:
          oneOf:
            - type: string
            - type: integer
        port:
          type: integer
        ranName:
          type: string
        setupFailure:
          $ref: '#/components/schemas/SetupFailure'
      additionalProperties: false
      type: object